wrapping file(REMOVE) commands inside if()

+++ support/cmake/packaging.cmake
@@ -387,14 +387,18 @@ if (UNIX AND NOT APPLE)
     # Static libraries
     file(GLOB_RECURSE _static_libs
       "$ENV{DESTDIR}${CMAKE_INSTALL_PREFIX}/*.a")
-    file(REMOVE ${_static_libs})
+    if (_static_libs)
+        file(REMOVE ${_static_libs})
+    endif ()
     
     # Git leftovers
     file(GLOB_RECURSE _gitfiles
       "$ENV{DESTDIR}${CMAKE_INSTALL_PREFIX}/*.git")
     file(GLOB_RECURSE _gitignores
       "$ENV{DESTDIR}${CMAKE_INSTALL_PREFIX}/.gitignore")
-    file(REMOVE ${_gitfiles} ${_gitignores})
+    if (_gitfiles OR _gitignores)
+      file(REMOVE ${_gitfiles} ${_gitignores})
+    endif ()
     
     # Source files
     file(GLOB_RECURSE _srcfiles
@@ -402,13 +406,17 @@ if (UNIX AND NOT APPLE)
       "$ENV{DESTDIR}${CMAKE_INSTALL_PREFIX}/*.cpp"
       "$ENV{DESTDIR}${CMAKE_INSTALL_PREFIX}/*.h"
       "$ENV{DESTDIR}${CMAKE_INSTALL_PREFIX}/*.hpp")
-    file(REMOVE ${_srcfiles})
+    if (_srcfiles)
+        file(REMOVE ${_srcfiles})
+    endif ()
     
     # CMake fragments
     file(GLOB_RECURSE _cmakefiles
       "$ENV{DESTDIR}${CMAKE_INSTALL_PREFIX}/*.cmake"
       "$ENV{DESTDIR}${CMAKE_INSTALL_PREFIX}/CMakeLists.txt")
-    file(REMOVE ${_cmakefiles})
+    if (_cmakefiles)
+        file(REMOVE ${_cmakefiles})
+    endif ()
     
     # Remove whole unwanted directories
     file(REMOVE_RECURSE
@@ -423,7 +431,9 @@ if (UNIX AND NOT APPLE)
     # zlib man page
     file(GLOB _zlib_man3
       "$ENV{DESTDIR}${CMAKE_INSTALL_PREFIX}/share/man/man3/zlib*")
-    file(REMOVE ${_zlib_man3})
+    if (_zlib_man3)
+        file(REMOVE ${_zlib_man3})
+    endif ()
 
     ### The above were not sufficient, so adding the lines below.
